Remove standardRule.UnitsComponent
  ref #31
