+++ src/standardMap/default/units-component.tsx
@@ -1,8 +1,8 @@
 import * as React from "react"
 import * as diplomacy from "js-diplomacy"
 
+import { standardRule } from "../../standardRule"
 import { Colors, UnitImage } from "./unit-image"
-import { UnitsComponent as BaseUnitsComponent } from "./../../standardRule/units-component"
 import { Point } from "../../util"
 
 import { provincePositionOf, locationPositionOf } from "./position"
@@ -23,6 +23,40 @@ export class UnitComponent extends UnitImage<diplomacy.standardMap.Power> {
   protected size = size
 }
 
-export class UnitsComponent extends BaseUnitsComponent<diplomacy.standardMap.Power> {
-  protected Unit = UnitComponent
+export class UnitsComponent
+  extends React.Component<standardRule.UnitsComponentProps<diplomacy.standardMap.Power>, {}> {
+  render () {
+    const units = Array.from(this.props.units)
+      .filter(unit => unit.status === null)
+      .map(elem => {
+        const unit = elem.unit
+        return <this.Unit
+          key={unit.toString()}
+          unit={elem}
+          on={(event) => {
+            if (this.props.on) {
+              this.props.on(event, unit)
+            }
+          }}/>
+      })
+    const dislodgedUnits = Array.from(this.props.units)
+      .filter(unit => unit.status !== null)
+      .map(elem => {
+        const unit = elem.unit
+        return <this.Unit
+          key={`${unit}-dislodged}`}
+          unit={elem}
+          on={(event) => {
+            if (this.props.on) {
+              this.props.on(event, unit)
+            }
+          }}/>
+      })
+
+    return <g>
+      {units}
+      {dislodgedUnits}
+    </g>
+  }
+  Unit = UnitComponent
 }
